+++ tests/test_scheduler.py
@@ -13,9 +13,9 @@ from scrapyd.utils import get_spider_queues
 def scheduler(tmpdir):
     eggs_dir = os.path.join(tmpdir, "eggs")
     dbs_dir = os.path.join(tmpdir, "dbs")
+    config = Config(values={"eggs_dir": eggs_dir, "dbs_dir": dbs_dir})
     os.makedirs(os.path.join(eggs_dir, "mybot1"))
     os.makedirs(os.path.join(eggs_dir, "mybot2"))
-    config = Config(values={"eggs_dir": eggs_dir, "dbs_dir": dbs_dir})
     return SpiderScheduler(config)
 
 
@@ -23,25 +23,31 @@ def test_interface(scheduler):
     verifyObject(ISpiderScheduler, scheduler)
 
 
-def test_list_update_projects(scheduler):
-    assert sorted(scheduler.list_projects()) == sorted(["mybot1", "mybot2"])
+# Need sorted(), because os.listdir() in FilesystemEggStorage.list_projects() uses an arbitrary order.
+def test_list_projects_update_projects(scheduler):
+    assert sorted(scheduler.list_projects()) == ["mybot1", "mybot2"]
 
     os.makedirs(os.path.join(scheduler.config.get("eggs_dir"), "mybot3"))
+
+    assert sorted(scheduler.list_projects()) == ["mybot1", "mybot2"]
+
     scheduler.update_projects()
 
-    assert sorted(scheduler.list_projects()) == sorted(["mybot1", "mybot2", "mybot3"])
+    assert sorted(scheduler.list_projects()) == ["mybot1", "mybot2", "mybot3"]
 
 
 def test_schedule(scheduler):
     queues = get_spider_queues(scheduler.config)
-    q1, q2 = queues["mybot1"], queues["mybot2"]
+    mybot1_queue = queues["mybot1"]
+    mybot2_queue = queues["mybot2"]
 
-    assert not q1.count()
+    assert not mybot1_queue.count()
+    assert not mybot2_queue.count()
 
     scheduler.schedule("mybot1", "myspider1", 2, a="b")
     scheduler.schedule("mybot2", "myspider2", 1, c="d")
     scheduler.schedule("mybot2", "myspider3", 10, e="f")
 
-    assert q1.pop() == {"name": "myspider1", "a": "b"}
-    assert q2.pop() == {"name": "myspider3", "e": "f"}
-    assert q2.pop() == {"name": "myspider2", "c": "d"}
+    assert mybot1_queue.pop() == {"name": "myspider1", "a": "b"}
+    assert mybot2_queue.pop() == {"name": "myspider3", "e": "f"}
+    assert mybot2_queue.pop() == {"name": "myspider2", "c": "d"}